% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_lists.R
\name{api_v60_url}
\alias{api_v60_url}
\title{Define API url}
\usage{
api_v60_url(server_name)
}
\arguments{
\item{server_name}{The server name as encoded in the url: `https//server_name.iformbuilder.com`}
}
\value{
A url for API requests, with \code{server_name} encoded
}
\description{
Use `api_v60_url` to define the API section of the request url.
If you do not have a dedicated server the server_name will be `app`
}
