% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_token.R
\name{base64url}
\alias{base64url}
\title{Encode to base64 using url-safe function. From httr}
\usage{
base64url(x)
}
\arguments{
\item{x}{A string to be encoded to base64}
}
\value{
The base64 encoded string
}
\description{
Encode to base64 using url-safe function. From httr
}
