% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_token.R
\name{base_url}
\alias{base_url}
\title{Define base_url}
\usage{
base_url(server_name)
}
\arguments{
\item{server_name}{The server name as encoded in the url: `https//server_name.iformbuilder.com`}
}
\value{
The base iFormBuilder url with \code{server_name} encoded
}
\description{
Use `base_url()` to define the base section of the request url.
If you do not have a dedicated server the server_name will be `app`
}
