% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{create_element}
\alias{create_element}
\title{Add element to page.}
\usage{
create_element(server_name, profile_id, access_token, page_id, name, label,
  description = "", data_type, data_size = 100, optionlist_id)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{access_token}{Access token produced by \code{iformr::get_iform_access_token}}

\item{page_id}{Page ID where the new element will be created.}

\item{name}{String of the element DCN. The provided name will be converted to be IFB database compliant (no special characters, all lowercase, spaces replaced with underscores.)}

\item{label}{Label for the element.}

\item{description}{Text description of the element.}

\item{data_type}{Integer indicating data type of the element. List of iFormBuilder data types: [https://iformbuilder.zendesk.com/hc/en-us/articles/201702880-Data-Types-and-Related-Details](https://iformbuilder.zendesk.com/hc/en-us/articles/201702880-Data-Types-and-Related-Details)}

\item{data_size}{*Optional* - length of the element; defaults to 100.}

\item{optionlist_id}{*Optional* - id of the option list to assign if a Select, Picklist, or Multi Picklist element is created.}
}
\value{
ID of the new element.
}
\description{
Adds a new element to a page.
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Create element
new_element <- create_element(
  server_name = "your_server_name",
  profile_id = "your_profile_id",
  access_token = access_token,
  page_id = "existing_page_id",
  name = "new_dcn_name",
  label = "new_element_label",
  description = "This is a new element.",
  data_type = 1
  }
}
