% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_lists.R
\name{create_new_option_list}
\alias{create_new_option_list}
\title{Create a new option list}
\usage{
create_new_option_list(server_name, profile_id, option_list_name, access_token)
}
\arguments{
\item{server_name}{The server name as encoded in the url: `https//server_name.iformbuilder.com`}

\item{profile_id}{The ID number of your profile}

\item{option_list_name}{A character name for the new option list}

\item{access_token}{The access_token required to establish communication with the API}
}
\value{
The ID of the new option list
}
\description{
Sends a request to the iFormBuilder API to create a new option list.
The new option list will be created with the name you supply.
}
