% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_records.R
\name{create_new_records}
\alias{create_new_records}
\title{Create new records}
\usage{
create_new_records(server_name, profile_id, page_id, access_token,
  record_data)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{page_id}{Integer ID of the page to insert new records.}

\item{access_token}{Access token produced by \code{iformr::get_iform_access_token}}

\item{record_data}{A dataframe containing the data to be added to the page.}
}
\value{
The created record ID or IDs.
}
\description{
Creates new records in an iFormBuilder page. The page must already
contain at least one record.
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")
}
}
\author{
Bill Devoe, \email{William.DeVoe@maine.gov}
}
