% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_records.R
\name{create_page}
\alias{create_page}
\title{Create page}
\usage{
create_page(server_name, profile_id, access_token, name, label)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{access_token}{Access token produced by \code{iformr::get_iform_access_token}}

\item{name}{String of new page name; coerced to iFormBuilder table
name conventions.}

\item{label}{String of the label for the new page.}
}
\value{
Integer of the new page ID.
}
\description{
Creates a new page in the given profile with the name and label specified.
The name provided will be converted to iFormBuilder standards; punctuation
and whitespace replaced with _ and all text to lowercase.
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Create new page
new_page_id <- create_page(
  server_name = "your_server_name",
  profile_id = "your_profile_id",
  access_token = access_token,
  name = "new_form_name",
  label = "New Form Label")
  }
}
\author{
Bill Devoe
}
