% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{data2form}
\alias{data2form}
\title{Create form from dataframe}
\usage{
data2form(server_name, profile_id, access_token, name, label, data)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{access_token}{Access token produced by \code{\link{get_iform_access_token}}}

\item{name}{String of new page name; coerced to iFormBuilder
table name conventions.}

\item{label}{String of the label for the new page.}

\item{data}{A dataframe whose structure will be used to
create the new form.}
}
\value{
The page ID of the created form.
}
\description{
Creates a form based on a dataframe. Dataframe classes are cast as
element types in the form.
}
\examples{
# Create a dataframe with some basic form fields
dat = tibble::tibble(survey_id = NA_integer_,
                     survey_datetime = as.POSIXct(NA, tz = "UTC"),
                     surveyor = NA_character_,
                     start_point = NA_real_,
                     fish_species = NA_integer_,
                     fish_count = NA_integer_,
                     end_point = NA_real_,
                     comment_text = NA_character_,
                     survey_completed = TRUE)

\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Create new form from dataframe
new_form <- data2form(
  server_name = "your_server_name",
  profile_id = "your_profile_id",
  access_token = access_token,
  name = "new_form_to_create",
  label = "New form based on an R dataframe",
  data = dat)
}

}
\author{
Bill DeVoe, \email{William.DeVoe@maine.gov}
}
