% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{delete_element}
\alias{delete_element}
\title{Delete an element}
\usage{
delete_element(server_name, profile_id, page_id, element_id, access_token)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{page_id}{Page ID from where the element will be deleted}

\item{element_id}{Integer ID of the new element that will be deleted}

\item{access_token}{Access token produced by \code{\link{get_iform_access_token}}}
}
\value{
ID of the deleted element.
}
\description{
Deletes an element from a page (form or subform)
}
\examples{
\dontrun{
# Pull out id from new_element created previously
new_element_id = new_element$id

# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Delete element
deleted_element <- delete_element(
  server_name = "your_server_name",
  profile_id = "your_profile_id",
  page_id = "existing_page_id",
  element_id = new_element_id,
  access_token = access_token)
  }

}
