% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_lists.R
\name{delete_option_list}
\alias{delete_option_list}
\title{Delete option list}
\usage{
delete_option_list(server_name, profile_id, access_token, option_list_id)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name}

\item{profile_id}{Integer of the iFormBuilder profile ID}

\item{access_token}{Access token produced by \code{\link{get_iform_access_token}}}

\item{option_list_id}{ID of the option list to be deleted.}
}
\value{
ID of the option list to be deleted.
}
\description{
Deletes an option list from a profile. Deleting options and option lists
should be done with consideration for existing data referencing the
list. As an alternative, options can be disabled by setting their condition
value to 'False'
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Delete option list
deleted_id <- delete_option_list(
  server_name = "your_server_name",
  profile_id = "your_profile_id",
  access_token = access_token,
  option_list_id)
  }
}
\author{
Bill DeVoe, \email{William.DeVoe@maine.gov}
}
