% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{delete_page}
\alias{delete_page}
\title{Delete page}
\usage{
delete_page(server_name, profile_id, page_id, access_token)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{page_id}{ID of the option list to be deleted.}

\item{access_token}{Access token produced by \code{\link{get_iform_access_token}}}
}
\value{
ID of the page that was deleted.
}
\description{
Deletes a single page (form or subform) from a profile. Use caution when
deleting forms. It should only be done with consideration for existing
data referencing the form.
}
\examples{
\dontrun{

# Pull out ID of page to delete
page_to_delete = forms_list$id[forms_list$name == "test_form_p"]

# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Delete option list
deleted_page_id <- delete_page(
  server_name = "your_server_name",
  profile_id = "your_profile_id",
  page_id,
  access_token = access_token)
}

}
