% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_records.R
\name{delete_record}
\alias{delete_record}
\title{Delete record}
\usage{
delete_record(server_name, profile_id, access_token, page_id, record_id)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{access_token}{Access token produced by \code{iformr::get_iform_access_token}}

\item{page_id}{Integer - ID of the page from which to delete the record.}

\item{record_id}{Integer - ID of the record to delete.}
}
\value{
ID of the record deleted.
}
\description{
Delete a single record.
}
\author{
Bill Devoe, \email{William.DeVoe@maine.gov}
}
