% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_records.R
\name{form_metadata}
\alias{form_metadata}
\title{Form metadata}
\usage{
form_metadata(server_name, profile_id, access_token, page_id, file,
  subforms = T)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{access_token}{Access token produced by \code{iformr::get_iform_access_token}}

\item{page_id}{ID of the form to get metadata from.}

\item{file}{Filename of the output Markdown file.}

\item{subforms}{**Optional** - Indicates if metadata should be generated for subforms.}
}
\description{
Builds a Markdown document containing metadata for a given
iFormBuilder form by querying the API for page and element
level information. By utilizing the description fields during
form building, detailed metadata can be built afterward using
this function.
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Add new data to form
sync_table(server_name, profile_id, access_token,
  data = "new_dataframe", form_name = "my_form",
  uid = "unique_id_col")
}
}
\author{
Bill Devoe, \email{William.DeVoe@maine.gov}
}
