% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{format_name}
\alias{format_name}
\title{Format a page or element name to be IFB compliant}
\usage{
format_name(name)
}
\arguments{
\item{name}{String of new page or element name.}
}
\value{
IFB compliant name.
}
\description{
Replaces whitespace and punctuation in an element/form name
with _ and converts name to lowercase.
Checks name against list of IFB reserved words, appending a '2'
after the name if it is in the reserved word list.
}
\author{
Bill Devoe, \email{William.DeVoe@maine.gov}
}
