% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_lists.R
\name{get_all_option_lists}
\alias{get_all_option_lists}
\title{Retrieve all option lists.}
\usage{
get_all_option_lists(server_name, profile_id, access_token)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{access_token}{Access token produced by \code{\link{get_iform_access_token}}}
}
\value{
Tibble of two columns containing the option list IDs and option
  list names: id <int>, name <chr>
}
\description{
Retrieves all option lists in a profile in chunks of 100 (API call limit).
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Get the id and name of all option lists in profile
option_lists <- get_all_option_lists(
  server_name = "your_server_name",
  profile_id = 123456,
  access_token = access_token)

# Inspect
option_lists
}
}
\author{
Bill Devoe, \email{William.DeVoe@maine.gov}
}
