% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_records.R
\name{get_all_pages_list}
\alias{get_all_pages_list}
\title{Get list of all pages (i.e., form, or subform) in a profile}
\usage{
get_all_pages_list(server_name, profile_id, access_token)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{access_token}{Access token produced by \code{iformr::get_iform_access_token}}
}
\value{
Tibble of two columns containing the page ID and page name: id <int>, name <chr>
}
\description{
Retrives a list of ALL pages in a profile, in chunks of 100 (limit). Returns a tibble with form id and form
name.
}
