% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_lists.R
\name{get_core_option_list_elements}
\alias{get_core_option_list_elements}
\title{Get core elements in an option list}
\usage{
get_core_option_list_elements(server_name, profile_id, optionlist_id,
  limit = 1000, offset = 0, access_token)
}
\arguments{
\item{server_name}{The server name as encoded in the url: `https//server_name.iformbuilder.com`}

\item{profile_id}{The ID number of your profile}

\item{optionlist_id}{The ID number for the option list}

\item{limit}{The maximum number of option list items to return}

\item{offset}{Skips the offset number of options before beginning to return}

\item{access_token}{The access_token required to establish communication with the API}
}
\value{
A dataframe of the core option list elements
}
\description{
Sends a request to the iFormBuilder API to return the core option list elements.
Function will return the id, sort_order, label, key_value, and condition_value.
}
