% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_lists.R
\name{get_option_list_id}
\alias{get_option_list_id}
\title{Get the ID of a single option list given an option list name}
\usage{
get_option_list_id(server_name, profile_id, option_list_name, limit = 1000,
  offset = 0, access_token)
}
\arguments{
\item{server_name}{The server name as encoded in the url: `https//server_name.iformbuilder.com`}

\item{profile_id}{The ID number of your profile}

\item{limit}{The maximum number of option lists to return}

\item{offset}{Skips the offset number of options before beginning to return}

\item{access_token}{The access_token required to establish communication with the API}
}
\value{
A listing of all option lists in the given profile
}
\description{
Sends a request to the iFormBuilder API to get the ID number for a single
option list. You only need to supply the name of the option list.
}
