% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_records.R
\name{get_page_id}
\alias{get_page_id}
\title{Get id of a single page (form) given a form name}
\usage{
get_page_id(server_name, profile_id, page_name, limit = 1000, offset = 0,
  access_token)
}
\arguments{
\item{server_name}{The server name as encoded in the url: `https//server_name.iformbuilder.com`}

\item{profile_id}{The ID number of your profile}

\item{page_name}{The name of the form or subform}

\item{limit}{The maximum number of form ids to return}

\item{offset}{Skips the offset number of ids before beginning to return}

\item{access_token}{The access_token required to establish communication with the API}
}
\value{
A dataframe listing the id and name of the given form
}
\description{
Sends a request to the iFormBuilder API to get a table containing
the id and name of a single form. You only need to supply the name
of the option list. Returns a dataframe with form id and form name.
}
