% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_records.R
\name{get_pages_list}
\alias{get_pages_list}
\title{Get list of all pages (forms, parent and subforms) in a profile}
\usage{
get_pages_list(server_name, profile_id, limit = 1000, offset = 0,
  access_token)
}
\arguments{
\item{server_name}{The server name as encoded in the url:
`https//server_name.iformbuilder.com`}

\item{profile_id}{The ID number of your profile}

\item{limit}{The maximum number of form ids to return}

\item{offset}{Skips the offset number of ids before beginning to return}

\item{access_token}{The access_token required to establish communication with
the API}
}
\value{
A dataframe of all forms (pages) in the given profile
}
\description{
Sends a request to the iFormBuilder API to get a listing of all forms and
subforms in the current profile. Returns a dataframe with form id and form
name.
}
