% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_photos}
\alias{get_photos}
\title{Get photos from form}
\usage{
get_photos(server_name, profile_id, access_token, page_id, photo, output,
  photoid)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{access_token}{Access token produced by \code{iformr::get_iform_access_token}}

\item{page_id}{Integer ID of the form to download photos from.}

\item{photo}{Character string of the photo element's DCN.}

\item{output}{Path of the output directory for the downloaded photos.}

\item{photoid}{Character string of the DCN to use as a filename
for the photo. Must contain unique values.}
}
\value{
Boolean True if succesful.
}
\description{
Downloads all of the photos taken with the photo element in a form
to a local directory, naming the files based on another field in the form.
Since the filenames produced in IFB are jibberish to people, the recommended practice
is to include a dynamically calculated element to uniquely identify each photo.
For example, in a workflow with a trip form, station subform, and station_photos
subform, the station_photos form would include a photoid field that created unique
photoids by concaternation of the trip ID, site ID, and index of the photo record.
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Download photos from a form

}
}
\author{
Bill Devoe, \email{William.DeVoe@maine.gov}
}
