% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_records.R
\name{get_selected_page_records}
\alias{get_selected_page_records}
\alias{data_feed_url}
\title{Get specific set of fields in a page (form)}
\usage{
get_selected_page_records(server_name, profile_id, page_id, fields = fields,
  limit = 100, offset = 0, access_token)

data_feed_url(server_name, parent_form_id, profile_id, parent_form_name,
  since_id, user_label, pw_label)
}
\arguments{
\item{server_name}{The server name as encoded in the url: `https//server_name.iformbuilder.com`}

\item{profile_id}{The ID number of your profile}

\item{page_id}{The id of the form or subform}

\item{fields}{A set of data fields (columns) to return}

\item{limit}{The maximum number of records to return}

\item{offset}{Skips the offset number of records before beginning to return}

\item{access_token}{The access_token required to establish communication with the API}

\item{parent_form_id}{The id of the parent form}

\item{parent_form_name}{The name of the parent form}

\item{since_id}{The record id indicating where to start downloading}

\item{user_label}{The name given to the username in the .Renviron file}

\item{pw_label}{Skips the offset number of records before beginning to return}

\item{server_name}{The server name as encoded in the url: `https//server_name.iformbuilder.com`}

\item{profile_id}{The ID number of your profile}
}
\value{
A dataframe of records from the specified fields

A .json file of records from the specified fields
}
\description{
Sends a request to the iFormBuilder API to get a list of all record
ids in a given form or subform.

Creates a url with username and password embedded that can be used to
download data using the data-feed mechanism instead of the API. In
general, this should be avoided, as the API mechanisms are much
safer. Returns a json file with all records since the specified
SINCE_ID.
}
