% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_token.R
\name{has_key}
\alias{has_key}
\title{Check if the application client_key was obtained}
\usage{
has_key(client_key)
}
\arguments{
\item{client_key}{The string value of the client_key}
}
\value{
TRUE if client_key was returned, else FALSE
}
\description{
Check if the application client_key was obtained
}
