% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_token.R
\name{has_secret}
\alias{has_secret}
\title{Check if application client_secret was obtained}
\usage{
has_secret(client_secret)
}
\arguments{
\item{client_secret}{The string value of the client_secret}
}
\value{
TRUE if client_secret was returned, else FALSE
}
\description{
Check if application client_secret was obtained
}
