% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_formats.R
\name{idate_time}
\alias{idate_time}
\title{Convert datetime values output by iFormBuilder}
\usage{
idate_time(dts, timezone = "America/Los_Angeles")
}
\arguments{
\item{dts}{A vector of datetimes as string values}

\item{timezone}{A character string in standard format specifying timezone.
Use \code{OlsonNames()} for a list of standard time zone character strings.}
}
\description{
Date and time values output from iFormBuilder can be difficult
  to work with. The output will typically look like:
  \code{"2014-11-13T15:04:00+00.00"}. The \code{idate_time()} function
  converts datetime outputs to standard format \code{"2014-10-13 08:04:03"}
  taking timezone into consideration.
}
\examples{
# Create vector that straddles change in daylight savings time
# Daylight savings time occurred: 2017-03-12 02:00:00
dts = c("2017-03-12T01:10:00+00:00", "", NA, "2017-03-12T02:10:00+00:00")

# Convert to standard format
idate_time(dts)
}
\seealso{
\code{\link{Sys.timezone}} for timezones
}
