% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_token.R, R/page_records.R
\name{iform_key}
\alias{iform_key}
\alias{iform_pw}
\title{Get application client_key from .Renviron file}
\usage{
iform_key(app_key_name)

iform_pw(pw_label)
}
\arguments{
\item{app_key_name}{The name given to the client_key in the .Renviron file}

\item{pw_label}{The name given to the password in the .Renviron file}
}
\value{
The API client_key for the application

The password for a read-only user in the given profile
}
\description{
Get application client_key from .Renviron file

Only needed to get password when downloading data using the data feed
mechanism. Avoid if possible. It is much safer to use the API.
}
