% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_token.R
\name{iform_secret}
\alias{iform_secret}
\title{Get application client_secret from .Renviron file}
\usage{
iform_secret(app_secret_name)
}
\arguments{
\item{app_secret_name}{The name given to the client_secret in the .Renviron file}
}
\value{
The API client_secret for the application
}
\description{
Get application client_secret from .Renviron file
}
