% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_records.R
\name{iform_user}
\alias{iform_user}
\title{Get username from the .Renviron file}
\usage{
iform_user(user_label)
}
\arguments{
\item{user_label}{The name given to the username in the .Renviron file}
}
\value{
The username for a read-only user in the given profile
}
\description{
Only needed to get username when downloading data using the data feed
mechanism. Avoid if possible. It is much safer to use the API.
}
