% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_formats.R
\name{itext_time}
\alias{itext_time}
\title{Convert datetime values output by iFormBuilder}
\usage{
itext_time(dts, create_tz = "America/Los_Angeles",
  timezone = "America/Los_Angeles")
}
\arguments{
\item{dts}{A vector of datetimes as string values}

\item{create_tz}{A character string in standard format specifying timezone.}

\item{timezone}{A character string in standard format specifying timezone.
Use \code{OlsonNames()} for a list of standard time zone character strings.}
}
\description{
Date and time values output from iFormBuilder using either the
  API functions or the data feed function can be difficult to work with. The
  output may look like: \code{"2014-11-13T15:04:00+00.00"} or: \code{Fri Apr
  08 2016 08:20:02 GMT-0700 (PDT)}, depending on which function is used. The
  \code{idate_time(), and itext_time()} functions convert these datetime
  outputs to standard format \code{"2014-10-13 08:04:03"}, taking timezone
  into consideration.
}
\examples{
# Create vector that straddles change in daylight savings time
dts = c("", NA, "Fri Apr 08 2016 08:20:02 GMT-0700 (PDT)",
        "Fri Feb 08 2016 08:20:02 GMT-0700 (PST)")

# Convert to standard format
itext_time(dts)
}
\seealso{
[timezones()]
}
