% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_token.R
\name{jencode}
\alias{jencode}
\title{Generate encoded request token}
\usage{
jencode(jheader, jpayload, client_secret)
}
\arguments{
\item{jheader}{The jwt_header in json format}

\item{jpayload}{The jwt_payload in json format}

\item{client_secret}{The client_secret assigned to the application}
}
\value{
The signed and encoded request token
}
\description{
Generate encoded request token
}
