% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_token.R
\name{jwt_payload}
\alias{jwt_payload}
\title{Define the jwt payload}
\usage{
jwt_payload(client_key, iat = NULL, exp = NULL, token_uri, duration = 600L)
}
\arguments{
\item{client_key}{The client_key assigned to the application}

\item{iat}{The issue time}

\item{exp}{The expiration time}

\item{token_uri}{The uri for requesting the access_token}

\item{duration}{The length of time between issue time and expiration time}
}
\value{
The json web token payload as a list
}
\description{
Define the jwt payload
}
