% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{rename_page}
\alias{rename_page}
\title{Rename page}
\usage{
rename_page(server_name, profile_id, access_token, page_id, name, label)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{access_token}{Access token produced by \code{\link{get_iform_access_token}}}

\item{page_id}{Integer of the page ID to rename.}

\item{name}{String of renamed page name; coerced to iFormBuilder
table name conventions.}

\item{label}{String of the renamed page label.}
}
\value{
Integer of the page ID.
}
\description{
Renames a page (form or subform) given a page_id and new name and label.
The name provided will be converted to iFormBuilder standards; punctuation
and whitespace will be replaced with _ and all text to lowercase.
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Rename page
rename_page_id <- rename_page(
  server_name = "your_server_name",
  profile_id = "your_profile_id",
  access_token = access_token,
  page_id = "existing_page_id",
  name = "new_page_name",
  label = "new_page_label")
  }
}
\author{
Bill DeVoe, \email{William.DeVoe@maine.gov}
}
