% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_license.R
\name{retrieve_all_device_licenses}
\alias{retrieve_all_device_licenses}
\title{Retrieve a list of device licenses}
\usage{
retrieve_all_device_licenses(server_name, profile_id, access_token)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{access_token}{Access token produced by \code{\link{get_iform_access_token}}}
}
\value{
Dataframe containing the ID, username, and device_id for
  all device licenses in a profile
}
\description{
Retrieves a list of all device licenses in a profile.
Core API call [Retrieve a List of Device Licenses](
https://iformbuilder.docs.apiary.io/#reference/device-license-resource/device-license-collection/retrieve-a-list-of-device-licenses)
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Retrieve a list of all device licenses in a profile
all_device_license_info <- retrieve_all_device_licenses(
  server_name = "your_server_name",
  profile_id = "your_profile_id",
  access_token = access_token)
}
}
