% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{retrieve_all_users}
\alias{retrieve_all_users}
\title{Retrieve a list of all users in a profile}
\usage{
retrieve_all_users(server_name, profile_id, access_token)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{access_token}{Access token produced by \code{\link{get_iform_access_token}}}
}
\value{
Dataframe containing all usernames and user IDs in the profile.
}
\description{
Retrieves a list of all users in a profile. Core API call
\href{https://iformbuilder.docs.apiary.io/#reference/user-resource/user-collection/retrieve-a-list-of-users}{Retrieve a List of Users}
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Get a list of all usernames and user IDs in a profile
all_users_info <- retrieve_all_users(
  server_name = "your_server_name",
  profile_id = 123456,
  access_token = access_token)

user_id = all_users_info$id[1]
}
}
