% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device_license.R
\name{retrieve_device_license}
\alias{retrieve_device_license}
\title{Retrieve a device license}
\usage{
retrieve_device_license(server_name, profile_id, access_token, license_id)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{access_token}{Access token produced by \code{\link{get_iform_access_token}}}

\item{license_id}{Integer id of the device license.}
}
\value{
List containing attributes of the device license
}
\description{
Retrieves a single device license. Core API call
\href{https://iformbuilder.docs.apiary.io/#reference/device-license-resource/device-license/retrieve-a-device-license}{Retrieve a Device License}
}
\examples{
\dontrun{

# Pull out the first device license
license_id = all_device_license_info$id[1]

# Retrieve details for a single device license
single_device_license_info <- retrieve_device_license(
  server_name = "your_server_name",
  profile_id = "your_profile_id",
  license_id = license_id,
  access_token = access_token)
}
}
\author{
Bill Devoe, \email{William.DeVoe@maine.gov}
}
