% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{retrieve_element_list}
\alias{retrieve_element_list}
\title{Retrieve a List of Elements}
\usage{
retrieve_element_list(
  server_name,
  profile_id,
  access_token,
  page_id,
  fields = "all"
)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{access_token}{Access token produced by \code{iformr::get_iform_access_token}}

\item{page_id}{Page ID where the new element will be created.}

\item{fields}{*Optional* - Defaults to 'all', which returns all fields for each
element. Optionally, a character vector of fields to return can be provided.}
}
\value{
Dataframe containing a row for each element and a column for each
  element field.
}
\description{
Retrieves a list of all the elements contained in a page.
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Get list of elements in form
elements <- retrieve_element_list(
  server_name = "your_server_name",
  profile_id = "your_profile_id",
  access_token = access_token,
  page_id = "existing_page_id",
  fields = 'all'
  }
}
\author{
Bill Devoe, \email{William.DeVoe@maine.gov}
}
