% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_records.R
\name{retrieve_page}
\alias{retrieve_page}
\title{Retrieve page}
\usage{
retrieve_page(server_name, profile_id, access_token, page_id)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{access_token}{Access token produced by \code{iformr::get_iform_access_token}}

\item{page_id}{ID of the page to retrieve.}
}
\value{
List containing page details.
}
\description{
Retrieves page details for a page ID.
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Retrieve page
page_data <- retrieve_page(
  server_name = "your_server_name",
  profile_id = "your_profile_id",
  access_token = access_token,
  page_id = page_id)
  }
}
\author{
Bill Devoe, \email{William.DeVoe@maine.gov}
}
