% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{retrieve_profile}
\alias{retrieve_profile}
\title{Retrieve a profile}
\usage{
retrieve_profile(server_name, profile_id, access_token)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{access_token}{Access token produced by \code{\link{get_iform_access_token}}}
}
\value{
List containing profile attributes.
}
\description{
Retrieves a profile. Core API call [Retrieve a Device License](
https://iformbuilder.docs.apiary.io/#reference/profile-resource/profile/retrieve-a-profile)
}
\author{
Bill Devoe, \email{William.DeVoe@maine.gov}
}
