% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sync_table}
\alias{sync_table}
\title{Sync table}
\usage{
sync_table(server_name, profile_id, access_token, data, form_name, label,
  uid, update = T, delete = F)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{access_token}{Access token produced by \code{iformr::get_iform_access_token}}

\item{data}{A dataframe containing the data to be synced with the page.}

\item{form_name}{The name of a page to sync the source data to; if the page does
not exist, it will be created.}

\item{label}{*Optional* - String of the label to be used if a new page is created.
If a label is not provided and a new page is created, the form_name argument will
be used to create a page label.}

\item{uid}{The name of the column in the source and IFB data that uniquely identifies
a record.}

\item{update}{*Optional* Defaults to True - If True, records in the form data
will be updated if the matching record in the source data is different.}

\item{delete}{*Optional* Defaults to False - If True, records in the form data
not present in the source data will be removed.}
}
\value{
The page ID of the existing or created form.
}
\description{
Syncs a dataframe with the contents of an IFB page. If the page does
not yet exist, it will be created and populated with the source data.
}
\details{
An example use case is syncing an existing database table with a
Smart Table Search form in IFB. All columns in the source data
with matching columns in the form data are synced. Data is synced
in a parent-to-child fashion:
- Rows in the source data not in the form data are added (based on
the unique identifier from the \code{uid} argument.)
- Rows in the form data that are not prseent in the source data
can be optionally removed using the \code{delete} argument.
- Rows in the form data are updated if any of their fields differ
from the matching row in the source data. Updating can be disabled
by passing False to the \code{update} argument.
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Add new data to form
sync_table(server_name, profile_id, access_token,
  data = "new_dataframe", form_name = "my_form",
  uid = "unique_id_col")
}
}
\author{
Bill Devoe, \email{William.DeVoe@maine.gov}
}
