% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_token.R
\name{token_url}
\alias{token_url}
\title{Define token_url}
\usage{
token_url(company)
}
\arguments{
\item{company}{The company name as encoded in the url}
}
\value{
A url for requesting an access_token with \code{company} encoded
}
\description{
Define token_url
}
