% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_records.R
\name{truncate_form}
\alias{truncate_form}
\title{Truncate form}
\usage{
truncate_form(server_name, profile_id, page_id, access_token)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name}

\item{profile_id}{Integer of the iFormBuilder profile ID}

\item{page_id}{Integer ID of the form to truncate.}

\item{access_token}{Access token produced by \code{\link{get_iform_access_token}}}
}
\value{
Integer vector of the deleted record IDs
}
\description{
Removes all records from a page, leaving the page structure.
USE WITH CAUTION!
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Get the id of a single form in the profile given the form name
form_id <- get_page_id(
  server_name = "your_server_name",
  profile_id = "your_profile_id",
  page_name = "your_form_p",
  access_token = access_token)

# Delete all records in the form
truncated_ids <- truncate_form(
  server_name = "your_server_name",
  profile_id = "your_profile_id",
  page_id = form_id,
  access_token = access_token)
}
}
\author{
Bill DeVoe, \email{William.DeVoe@maine.gov}
}
