% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{update_option_lists}
\alias{update_option_lists}
\title{Update iFormBuilder option lists}
\usage{
update_option_lists(access_token, server_name, profile_id, option_lists)
}
\arguments{
\item{access_token}{Access token produced by
\code{\link[iformr]{get_iform_access_token}}}

\item{server_name}{String of the iFormBuilder server name}

\item{profile_id}{Integer of the iFormBuilder profile ID}

\item{option_lists}{A dataframe containing option lists withname, key_value,
label, condition_value, and sort_order.}
}
\value{
No return value.
}
\description{
For an input dataframe of option lists, adds new option lists
to the profile, adds new options to existing option lists, and updates
condition_value for existing options.
}
