% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_records.R
\name{update_records}
\alias{update_records}
\title{Update records (up to 100)}
\usage{
update_records(server_name, profile_id, page_id, access_token, record_data)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{page_id}{Integer ID of the page to perform the update on.}

\item{access_token}{Access token produced by
\code{\link{get_iform_access_token}}}

\item{record_data}{A dataframe containing the update data.The dataframe must
contain columns for the fields to update, along with an 'id' column
containing the record IDs to update.}
}
\value{
The updated record ID or IDs.
}
\description{
Updates up to 100 records in an iFormBuilder page.
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")
}
}
\author{
Bill Devoe, \email{William.DeVoe@maine.gov}
}
