% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_lists.R
\name{add_options_to_list}
\alias{add_options_to_list}
\title{Add option values to new option list}
\usage{
add_options_to_list(server_name, profile_id, optionlist_id, option_values,
  access_token)
}
\arguments{
\item{server_name}{The server name as encoded in the url: `https//server_name.iformbuilder.com`}

\item{profile_id}{The ID number of your profile}

\item{optionlist_id}{The ID number for the option list}

\item{option_values}{A list of option values in json format}

\item{access_token}{The access_token required to establish communication with the API}
}
\value{
A vector of option list element IDs, one for each option in the list
}
\description{
Sends a request to the iFormBuilder API to append a list of options
in json format to an existing option list. Make sure that all options
list key_values are unique or the list will not be posted.
}
