% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_token.R
\name{base64url}
\alias{base64url}
\title{Encode to url-safe base64}
\usage{
base64url(x)
}
\arguments{
\item{x}{A string to be encoded to base64}
}
\value{
The base64 encoded string
}
\description{
Taken from the httr package. Will be replaced in
a future version by importing the base64url package
}
