% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_token.R
\name{base_url}
\alias{base_url}
\title{Define base_url}
\usage{
base_url(company)
}
\arguments{
\item{company}{The company name as encoded in the url}
}
\value{
The base iFormBuilder url with \code{company} encoded
}
\description{
Use `base_url()` to define the base section of the request url.
For example: `https://app.iformbuilder.com`.
}
