% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_records.R
\name{copy_page}
\alias{copy_page}
\title{Copy page}
\usage{
copy_page(server_name, profile_id, access_token, page_id)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{access_token}{Access token produced by \code{iformr::get_iform_access_token}}

\item{page_id}{Integer of the page ID to copy.}
}
\value{
Integer of the new page ID.
}
\description{
Copies a page to a new page in the profile.
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Copy page
new_page_id <- copy_page(
  server_name = "your_server_name",
  profile_id = "your_profile_id",
  access_token = access_token,
  page_id = "existing_page_id"
  }

}
