% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_records.R
\name{data_feed_url}
\alias{data_feed_url}
\title{Compose a url to get data via the data feed mechanism}
\usage{
data_feed_url(server_name, parent_form_id, profile_id, parent_form_name,
  since_id, user_label, pw_label)
}
\arguments{
\item{server_name}{The server name as encoded in the url:
`https//server_name.iformbuilder.com`}

\item{parent_form_id}{The id of the parent form}

\item{profile_id}{The id number of your profile}

\item{parent_form_name}{The name of the parent form}

\item{since_id}{The record id indicating where to start downloading}

\item{user_label}{The name given to the username in the .Renviron file}

\item{pw_label}{Skips the offset number of records before beginning to return}
}
\value{
A url that can be used to request form data
}
\description{
Creates a url with username and password embedded that can be used to
download data using the data-feed mechanism instead of the API. In general,
this should be avoided, as the API mechanisms are much safer. Returns a json
file with all records submitted since the specified \code{since_id}.
}
\examples{
\dontrun{
# Generate a url to retrieve data via the data feed mechanism
url <- data_feed_url(
  server_name = "your_server_name",
  parent_form_id = 456789,
  profile_id = 123456,
  parent_form_name = "spawning_ground_p",
  since_id = 3,
  user_label = "your_user_label",
  pw_label = "your_pw_label")

# Retrieve the form data into an R list
form_data <- jsonlite::fromJSON(url)
}
}
