% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_lists.R
\name{delete_options_in_list}
\alias{delete_options_in_list}
\title{Delete all or some options in an option list}
\usage{
delete_options_in_list(server_name, profile_id, optionlist_id,
  fields = "fields", id_values, limit = 1000, offset = 0, access_token)
}
\arguments{
\item{server_name}{The server name as encoded in the url: `https//server_name.iformbuilder.com`}

\item{profile_id}{The ID number of your profile}

\item{optionlist_id}{The ID number for the option list}

\item{fields}{Placeholder for fields to delete, not yet implemented}

\item{id_values}{A .json list of ids for elements to delete}

\item{limit}{The maximum number of option elements to delete}

\item{offset}{Skips the offset number of options before beginning to delete}

\item{access_token}{The access_token required to establish communication with the API}
}
\value{
A vector of option list element IDs, one for each option in the list
}
\description{
Sends a request to the iFormBuilder API to delete a list of option elements.
The elements to delete are specified by a .json list of element ids. Sort
order will automatically be reassigned after deleting specified elements.
}
