% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_records.R
\name{delete_record}
\alias{delete_record}
\title{Delete record}
\usage{
delete_record(server_name, profile_id, page_id, record_id, access_token)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name}

\item{profile_id}{Integer of the iFormBuilder profile ID}

\item{page_id}{Integer ID of the page from which to delete the record.}

\item{record_id}{Integer ID of the record to delete.}

\item{access_token}{Access token produced by \code{\link{get_iform_access_token}}}
}
\value{
ID of the record deleted.
}
\description{
Delete a single record
}
\examples{
\dontrun{
# Identify the last record added to the form so it can be deleted
record_to_delete = max(parent_form_records$id)

# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Delete the last record that was added
 deleted_record = delete_record(
   server_name = "your_server_name",
   profile_id = "your_profile_id",
   page_id = form_id,
   record_id = record_to_delete,
   access_token = access_token)
}
}
\author{
Bill DeVoe, \email{William.DeVoe@maine.gov}
}
