% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{form_metadata}
\alias{form_metadata}
\title{Form metadata}
\usage{
form_metadata(server_name, profile_id, access_token, page_id, filename,
  subforms = T, sub = F)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{access_token}{Access token produced by \code{\link{get_iform_access_token}}}

\item{page_id}{ID of the form to get metadata from.}

\item{filename}{Filename of the output Markdown file.}

\item{subforms}{**Optional** - Indicates if metadata should be generated for subforms.
Defaults to True.}

\item{sub}{**Optional** - Defaults to False. Used by function to self-reference
and append subform metadata to beginning file.}
}
\value{
Add this later.
}
\description{
Builds a Markdown document containing metadata for a given
iFormBuilder form by querying the API for page and element
level information. By utilizing the description fields during
form building, detailed metadata can be built afterward using
this function.
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Create metadata for form.
form_metadata(server_name, profile_id, access_token,
                page_id = 012345, filename = "metadata", subforms = T)
}
}
\author{
Bill Devoe, \email{William.DeVoe@maine.gov}
}
