% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_token.R
\name{get_iform_access_token}
\alias{get_iform_access_token}
\title{Request an access_token}
\usage{
get_iform_access_token(server_name, app_key_name, app_secret_name)
}
\arguments{
\item{server_name}{The server name as encoded in the url: `https//server_name.iformbuilder.com`}

\item{app_key_name}{The name given to the client_key in the .Renviron file}

\item{app_secret_name}{The name given to the client_secret in the .Renviron file}
}
\value{
An access_token that expires after ten minutes
}
\description{
Sends a request to iFormBuilder for an access_token. This is needed
in order to authorize communication with the iFormBuilder API.
If you do not have a dedicated server your server_name will be `app`.
For the Washington Dept of Fish and Wildlife, server_name is `wdfw`.
}
