% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_token.R
\name{get_iform_access_token}
\alias{get_iform_access_token}
\title{Sends a request to iFormBuilder for an access_token. This is needed
in order to authorize communication with the iFormBuilder API}
\usage{
get_iform_access_token(company_name, app_key_name, app_secret_name)
}
\arguments{
\item{company_name}{The company name as encoded in the url}

\item{app_key_name}{The name given to the client_key in the .Renviron file}

\item{app_secret_name}{The name given to the client_secret in the .Renviron file}
}
\value{
An access_token that expires after ten minutes
}
\description{
If you do not have a dedicated server your company_name will be `app`.
For the Washington Dept of Fish and Wildlife, company_name is `wdfw`.
}
