% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_lists.R
\name{get_option_list_element_ids}
\alias{get_option_list_element_ids}
\title{Get list of option_ids for a given element}
\usage{
get_option_list_element_ids(server_name, profile_id, optionlist_id, element,
  limit = 1000, offset = 0, access_token)
}
\arguments{
\item{server_name}{The server name as encoded in the url:
`https//server_name.iformbuilder.com`}

\item{profile_id}{The ID number of your profile}

\item{optionlist_id}{The ID number for the option list}

\item{element}{The specific option list element. For example:
"condition_value".}

\item{limit}{The maximum number of option element IDs to return}

\item{offset}{Skips the offset number of options before beginning to return}

\item{access_token}{The access_token required to establish communication with
the API}
}
\value{
A vector of option list element IDs for the element selected.
}
\description{
Sends a request to the iFormBuilder API to get a list of all element IDs in
an option list for a specific field. For example: condition_value.
}
