% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_records.R
\name{get_page_id}
\alias{get_page_id}
\title{Get id of a single page (i.e., form, or subform) given a form name}
\usage{
get_page_id(server_name, profile_id, page_name, limit = 1000, offset = 0,
  access_token)
}
\arguments{
\item{server_name}{The server name as encoded in the url:
`https//server_name.iformbuilder.com`}

\item{profile_id}{The id number of your profile}

\item{page_name}{The name of the form or subform}

\item{limit}{The maximum number of form ids to return}

\item{offset}{Skips the offset number of ids before beginning to return}

\item{access_token}{The access_token required to establish communication with
the API}
}
\value{
An integer id for the given form
}
\description{
Sends a request to the iFormBuilder API to get the id number of a single
form. You only need to supply the name of the option list. Returns an integer
id for the given form.
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Get the id of a single form in the profile given the form name
form_id <- get_page_id(
  server_name = "your_server_name",
  profile_id = 123456,
  page_name = "spawning_ground_p",
  access_token = access_token)

# Inspect the form_id
form_id
}
}
