% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_records.R
\name{get_page_record_list}
\alias{get_page_record_list}
\title{Get list of all record ids in a single page (i.e., form, or subform)}
\usage{
get_page_record_list(
  server_name,
  profile_id,
  page_id,
  limit = 1000,
  offset = 0,
  access_token
)
}
\arguments{
\item{server_name}{The server name as encoded in the url:
`https//server_name.iformbuilder.com`}

\item{profile_id}{The id number of your profile}

\item{page_id}{The id number of the form}

\item{limit}{The maximum number of form ids to return}

\item{offset}{Skips the offset number of ids before beginning to return}

\item{access_token}{The access_token required to establish communication with
the API}
}
\value{
A vector of record ids from the given form
}
\description{
Sends a request to the iFormBuilder API to get a list of all record ids in a
given form or subform.
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Get the id of a single form in the profile given the form name
form_id <- get_page_id(
  server_name = "your_server_name",
  profile_id = 123456,
  page_name = "your_form_p",
  access_token = access_token)

# Get a list of all record ids in the specified form
record_ids <- get_page_record_list(
  server_name = "your_server_name",
  profile_id = 123456,
  page_id = form_id,
  access_token = access_token)

# Inspect the top five record_ids
head(record_ids, 5)
}
}
