% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_records.R
\name{get_pages_list}
\alias{get_pages_list}
\title{Get list of all pages (i.e., form, or subform) in a profile}
\usage{
get_pages_list(server_name, profile_id, limit = 1000, offset = 0,
  access_token)
}
\arguments{
\item{server_name}{The server name as encoded in the url:
`https//server_name.iformbuilder.com`}

\item{profile_id}{The id number of your profile}

\item{limit}{The maximum number of form ids to return}

\item{offset}{Skips the offset number of ids before beginning to return}

\item{access_token}{The access_token required to establish communication with
the API}
}
\value{
A dataframe of all forms (pages) in the given profile
}
\description{
Sends a request to the iFormBuilder API to get a listing of all forms and
subforms in the current profile. Returns a dataframe with form id and form
name. Limited to 100 records per API call.
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Get the id and name of all forms in profile
forms_list <- get_pages_list(
  server_name = "your_server_name",
  profile_id = 123456,
  access_token = access_token)

# Inspect
forms_list
}
}
