% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_records.R
\name{get_selected_page_records}
\alias{get_selected_page_records}
\title{Get specific set of fields in a page (form)}
\usage{
get_selected_page_records(server_name, profile_id, page_id, fields = fields,
  limit = 100, offset = 0, access_token)
}
\arguments{
\item{server_name}{The server name as encoded in the url:
`https//server_name.iformbuilder.com`}

\item{profile_id}{The ID number of your profile}

\item{page_id}{The id of the form or subform}

\item{fields}{A set of data fields (columns) to return}

\item{limit}{The maximum number of records to return}

\item{offset}{Skips the offset number of records before beginning to return}

\item{access_token}{The access_token required to establish communication with
the API}
}
\value{
A dataframe of records from the specified fields
}
\description{
Sends a request to the iFormBuilder API to get a list of all record ids in a
given form or subform.
}
