% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_token.R
\name{hex_to_raw}
\alias{hex_to_raw}
\title{Get raw output from sha256() signing. From openssl}
\usage{
hex_to_raw(str)
}
\arguments{
\item{str}{A string to be converted from hex to raw format}
}
\value{
The raw string
}
\description{
Get raw output from sha256() signing. From openssl
}
