% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_token.R
\name{iform_key}
\alias{iform_key}
\title{Get application client_key from .Renviron file}
\usage{
iform_key(app_key_name)
}
\arguments{
\item{app_key_name}{The name given to the client_key in the .Renviron file}
}
\value{
The API client_key for the application
}
\description{
Get application client_key from .Renviron file
}
