% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{retrieve_company_info}
\alias{retrieve_company_info}
\title{Retrieve company info}
\usage{
retrieve_company_info(server_name, profile_id, access_token)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name}

\item{profile_id}{Integer of the iFormBuilder profile ID}

\item{access_token}{Access token produced by \code{\link{get_iform_access_token}}}
}
\value{
Dataframe containing company info attributes
}
\description{
Retrieves company (profile) info. Core API call [Retrieve Company Info](
https://iformbuilder.docs.apiary.io/#reference/profile-resource/company-info/retrieve-company-info)
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Get a list of all usernames and user IDs in a profile
company_info <- retrieve_company_info(
  server_name = "your_server_name",
  profile_id = 123456,
  access_token = access_token)
}
}
\author{
Bill Devoe, \email{William.DeVoe@maine.gov}
}
