% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{retrieve_user}
\alias{retrieve_user}
\title{Retrieve a user}
\usage{
retrieve_user(server_name, profile_id, user_id, access_token)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{user_id}{Integer of the user ID to retrieve.}

\item{access_token}{Access token produced by \code{\link{get_iform_access_token}}}
}
\value{
List containing the user details.

#' # Get a list of details for a given user
all_users_info <- retrieve_all_users(
  server_name = "your_server_name",
  profile_id = 123456,
  user_id = user_id,
  access_token = access_token)
}
\description{
Retrieves information for a specific user. Core API call [Retrieve a User](
https://iformbuilder.docs.apiary.io/#reference/user-resource/user/retrieve-a-user)
}
\author{
Bill Devoe, \email{William.DeVoe@maine.gov}
}
