% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notifications.R
\name{send_notifications}
\alias{send_notifications}
\title{Send notifications}
\usage{
send_notifications(server_name, profile_id, access_token, message, users)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{access_token}{Access token produced by \code{\link{get_iform_access_token}}}

\item{message}{String containing the message to send.}

\item{users}{Numerical vector containing one or more user IDs,
ie c(1234, 3456)}
}
\value{
Vector of the user IDs that were successfully notified.
}
\description{
Sends a notification to a user or users. Core API call
\href{https://iformbuilder.docs.apiary.io/#reference/notification-resource/notification/send-notifications}{Send Notifications}
}
\examples{
\dontrun{
# Get access_token
access_token <- get_iform_access_token(
  server_name = "your_server_name",
  client_key_name = "your_client_key_name",
  client_secret_name = "your_client_secret_name")

# Get a list of all usernames and user IDs in a profile
all_users_info <- retrieve_all_users(
  server_name = "your_server_name",
  profile_id = 123456,
  access_token = access_token)

# Pull out first user_id in the list
user_ids = all_users_info$id

# Send a notification
notified_user_ids <- send_notifications(
  server_name = "your_server_name",
  profile_id = 123456,
  message = "Please sync your device",
  users = user_ids,
  access_token = access_token)
}
}
\author{
Bill DeVoe, \email{William.DeVoe@maine.gov}
}
