% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_token.R
\name{token_url}
\alias{token_url}
\title{Define token_url}
\usage{
token_url(server_name)
}
\arguments{
\item{server_name}{The server name as encoded in the url: `https//server_name.iformbuilder.com`}
}
\value{
A url for requesting an access_token with \code{server_name} encoded
}
\description{
Define token_url
}
