% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{truncate_form}
\alias{truncate_form}
\title{Truncate form}
\usage{
truncate_form(server_name, profile_id, access_token, page_id)
}
\arguments{
\item{server_name}{String of the iFormBuilder server name.}

\item{profile_id}{Integer of the iFormBuilder profile ID.}

\item{access_token}{Access token produced by \code{iformr::get_iform_access_token}}

\item{page_id}{Integer ID of the form to truncate.}
}
\value{
Boolean True if succesful.
}
\description{
Removes all records from a page, leaving the page structure.
USE WITH CAUTION!
}
\author{
Bill Devoe, \email{William.DeVoe@maine.gov}
}
