% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_lists.R
\name{update_options_in_list}
\alias{update_options_in_list}
\title{Update values in an existing option list}
\usage{
update_options_in_list(server_name, profile_id, optionlist_id, option_values,
  fields = fields, limit = 1000, offset = 0,
  access_token = access_token)
}
\arguments{
\item{server_name}{The server name as encoded in the url: `https//server_name.iformbuilder.com`}

\item{profile_id}{The ID number of your profile}

\item{optionlist_id}{The ID number for the option list}

\item{option_values}{A json object containing new option list values}

\item{fields}{A list of the specific fields to update}

\item{limit}{The maximum number of option list items to return}

\item{offset}{Skips the offset number of options before beginning to update}

\item{access_token}{The access_token required to establish communication with the API}
}
\value{
A vector of option IDs for elements that were updated
}
\description{
Sends a request to the iFormBuilder API to update an option list. Option values
for the specified fields will be updated to the new values supplied in the
json object `option_values`.
}
